Revert "Cleanup and organize code (#6)"
This reverts commit f68e2b44e9b40422150a01afb6597abf1c1972a3.
