+++ src/EzActiveDirectory/EzActiveDirectory.csproj
@@ -1,10 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.SDK">
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>netstandard2.0</TargetFramework>
     <LangVersion>Latest</LangVersion>
   </PropertyGroup>
-	
   <ItemGroup>
-    <PackageReference Include="System.DirectoryServices" Version="8.0.0" />
+    <PackageReference Include="System.DirectoryServices" Version="6.0.0" />
   </ItemGroup>
 </Project>
