--- src/EzActiveDirectory/Models/Computer.cs
@@ -1,6 +1,0 @@
-﻿namespace EzActiveDirectory.Models;
-public class Computer
-{
-    public string Name { get; set; }
-    public string Path { get; set; }
-}
